+++ packages/orm/angel_orm_mysql/lib/src/orm_mysql.dart
@@ -4,12 +4,14 @@ import 'package:logging/logging.dart';
 import 'package:mysql_client/mysql_client.dart';
 
 class MySqlExecutor extends QueryExecutor {
-  /// An optional [Logger] to write to.
-  final Logger? logger;
+  /// An optional [Logger] to write to. A default logger will be used if not set
+  late Logger logger;
 
   final MySQLConnection _connection;
 
-  MySqlExecutor(this._connection, {this.logger});
+  MySqlExecutor(this._connection, {Logger? logger}) {
+    this.logger = logger ?? Logger('MySqlExecutor');
+  }
 
   final Dialect _dialect = const MySQLDialect();
 
@@ -18,14 +20,9 @@ class MySqlExecutor extends QueryExecutor {
 
   Future<void> close() {
     return _connection.close();
-    /*
-    if (_connection is MySqlConnection) {
-      return (_connection as MySqlConnection).close();
-    } else {
-      return Future.value();
-    }
-    */
   }
+
+  MySQLConnection get rawConnection => _connection;
 
   /*
   Future<Transaction> _startTransaction() {
@@ -86,13 +83,23 @@ class MySqlExecutor extends QueryExecutor {
     // Change @id -> ?
     for (var name in substitutionValues.keys) {
       query = query.replaceAll('@$name', ':$name');
+
+      // Convert UTC time to local time
+      var value = substitutionValues[name];
+      if (value is DateTime && value.isUtc) {
+        var t = value.toLocal();
+        logger.warning('Datetime deteted: $name');
+        logger.warning('Datetime: UTC -> $value, Local -> $t');
+
+        substitutionValues[name] = t;
+      }
     }
 
     //var params = substitutionValues.values.toList();
-    var params = [];
+    //var params = [];
 
-    logger?.warning('Query: $query');
-    logger?.warning('Values: $substitutionValues');
+    logger.warning('Query: $query');
+    logger.warning('Values: $substitutionValues');
     //logger?.warning('Returning Query: $returningQuery');
 
     if (returningQuery.isNotEmpty) {
@@ -101,53 +108,64 @@ class MySqlExecutor extends QueryExecutor {
       if (query.startsWith("INSERT")) {
         var result = await _connection.execute(query, substitutionValues);
 
-        logger?.warning(result.lastInsertID);
+        logger.warning(result.lastInsertID);
 
         query = returningQuery;
         //logger?.warning('Result.insertId: ${result.insertId}');
         // Has primary key
-        //if (result.insertId != 0) {
         if (returningQuery.endsWith('.id=?')) {
           query = query.replaceAll("?", ":id");
-          //params = [result.lastInsertID];
           substitutionValues.clear();
           substitutionValues['id'] = result.lastInsertID;
+        } else {
+          query = _convertSQL(query, substitutionValues);
         }
       } else if (query.startsWith("UPDATE")) {
         await _connection.execute(query, substitutionValues);
         query = returningQuery;
-        params = [];
       }
     }
 
-    logger?.warning('Query 2: $query');
-    logger?.warning('Values 2: $substitutionValues');
+    logger.warning('Query 2: $query');
+    logger.warning('Values 2: $substitutionValues');
 
     // Handle select
     return _connection.execute(query, substitutionValues).then((results) {
-      logger?.warning("SELECT");
-      //for (var element in results.cols) {
-      //  print(
-      //      '${element.name} ${element.type.toString()} ${element.runtimeType.toString()}');
-      //}
+      logger.warning("SELECT");
 
       return results.rows.map((r) => r.typedAssoc().values.toList()).toList();
     });
   }
 
-  @override
-  Future<T> transaction<T>(FutureOr<T> Function(QueryExecutor) f) async {
-    logger?.warning("Transaction");
-    return f(this);
-    /*
-    if (_connection is! MySqlConnection) {
-      return await f(this);
+  String _convertSQL(String query, Map<String, dynamic> substitutionValues) {
+    var newQuery = query;
+    for (var k in substitutionValues.keys) {
+      var fromPattern = '.$k = ?';
+      var toPattern = '.$k = :$k';
+      newQuery = newQuery.replaceFirst(fromPattern, toPattern);
     }
 
-    await _connection.transaction((context) async {
-      var executor = MySqlExecutor(context, logger: logger);
+    return newQuery;
+  }
+
+  @override
+  Future<T> transaction<T>(FutureOr<T> Function(QueryExecutor) f) async {
+    logger.warning("Transaction");
+
+    T? returnValue = await _connection.transactional((ctx) async {
+      try {
+        logger.fine('Entering transaction');
+        var tx = MySqlExecutor(ctx, logger: logger);
+        return await f(tx);
+      } catch (e) {
+        logger.severe('Failed to run transaction', e);
+        rethrow;
+      } finally {
+        logger.fine('Exiting transaction');
+      }
     });
-    */
+
+    return returnValue!;
   }
   /*
   @override