+++ packages/orm/angel_orm_mysql/test/migrations/author.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE authors (
+CREATE TABLE IF NOT EXISTS authors (
     id serial PRIMARY KEY,
     name varchar(255) UNIQUE NOT NULL,
     created_at timestamp,