+++ packages/orm/angel_orm_mysql/test/migrations/has_car.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE has_cars (
+CREATE TABLE IF NOT EXISTS has_cars (
     id serial PRIMARY KEY,
     type int not null,
     created_at timestamp,