+++ packages/orm/angel_orm_mysql/test/migrations/has_map.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE has_maps (
+CREATE TABLE IF NOT EXISTS has_maps (
     id serial PRIMARY KEY,
     value jsonb not null,
     list jsonb not null,