+++ packages/orm/angel_orm_mysql/test/migrations/role.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE roles (
+CREATE TABLE IF NOT EXISTS roles (
   id serial PRIMARY KEY,
   name varchar(255),
   created_at timestamp,