+++ packages/orm/angel_orm_mysql/test/migrations/user_role.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE role_users (
+CREATE TABLE IF NOT EXISTS role_users (
   id serial PRIMARY KEY,
   user_id int NOT NULL,
   role_id int NOT NULL,