+++ packages/orm/angel_orm_test/lib/src/models/leg.g.dart
@@ -111,8 +111,8 @@ class LegQuery extends Query<Leg, LegQueryWhere> {
     }
     var model = Leg(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         name: fields.contains('name') ? (row[3] as String?) : null);
     if (row.length > 4) {
       var modelOpt = FootQuery().parseRow(row.skip(4).take(5).toList());
@@ -240,12 +240,10 @@ class FootQuery extends Query<Foot, FootQueryWhere> {
     }
     var model = Foot(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         legId: fields.contains('leg_id') ? (row[3] as int?) : null,
-        nToes: fields.contains('n_toes')
-            ? double.tryParse(row[4].toString())
-            : null);
+        nToes: fields.contains('n_toes') ? mapToDouble(row[4]) : null);
     return Optional.of(model);
   }