+++ packages/orm/angel_orm_test/lib/src/models/person_order.g.dart
@@ -93,13 +93,11 @@ class PersonOrderQuery extends Query<PersonOrder, PersonOrderQueryWhere> {
     }
     var model = PersonOrder(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         personId: fields.contains('person_id') ? (row[3] as int?) : null,
         name: fields.contains('name') ? (row[4] as String?) : null,
-        price: fields.contains('price')
-            ? double.tryParse(row[5].toString())
-            : null,
+        price: fields.contains('price') ? mapToDouble(row[5]) : null,
         deleted: fields.contains('deleted') ? mapToBool(row[6]) : null);
     return Optional.of(model);
   }
@@ -258,12 +256,10 @@ class OrderWithPersonInfoQuery
     }
     var model = OrderWithPersonInfo(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         name: fields.contains('name') ? (row[3] as String?) : null,
-        price: fields.contains('price')
-            ? double.tryParse(row[4].toString())
-            : null,
+        price: fields.contains('price') ? mapToDouble(row[4]) : null,
         deleted: fields.contains('deleted') ? mapToBool(row[5]) : null,
         personName: fields.contains('person_name') ? (row[6] as String?) : null,
         personAge: fields.contains('person_age') ? (row[7] as int?) : null);