+++ packages/orm/angel_orm_test/lib/src/models/user.g.dart
@@ -127,8 +127,8 @@ class UserQuery extends Query<User, UserQueryWhere> {
     }
     var model = User(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         username: fields.contains('username') ? (row[3] as String?) : null,
         password: fields.contains('password') ? (row[4] as String?) : null,
         email: fields.contains('email') ? (row[5] as String?) : null);
@@ -487,8 +487,8 @@ class RoleQuery extends Query<Role, RoleQueryWhere> {
     }
     var model = Role(
         id: fields.contains('id') ? row[0].toString() : null,
-        createdAt: fields.contains('created_at') ? (row[1] as DateTime?) : null,
-        updatedAt: fields.contains('updated_at') ? (row[2] as DateTime?) : null,
+        createdAt: fields.contains('created_at') ? mapToDateTime(row[1]) : null,
+        updatedAt: fields.contains('updated_at') ? mapToDateTime(row[2]) : null,
         name: fields.contains('name') ? (row[3] as String?) : null);
     if (row.length > 4) {
       var modelOpt = UserQuery().parseRow(row.skip(4).take(6).toList());